backend/requirements.txt
asgiref==3.8.0
Django==5.0.3
pillow==10.2.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.10.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-ckeditor